feat: :sparkles: Added CTA section with signup form
Added the CTA (Call to Action) section for the Crafts Shop Website Design.

This section encourages users to sign up and get 10% off.

It includes a headline, a descriptive paragraph, and a signup form where users can enter their email and subscribe.

The form consists of an email input field and a submit button.